requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework~=3.15.2
drf-spectacular~=0.28.0